refactor some utils into ingestion/common
