+++ src/encoded/ingestion_listener.py
@@ -29,7 +29,7 @@ from pyramid.view import view_config
 from snovault.util import debug_log
 from vcf import Reader
 from .commands.ingest_vcf import VCFParser
-from .ingestion.common import register_path_content_type, metadata_bundles_bucket, get_parameter
+from .ingestion.common import register_path_content_type, metadata_bundles_bucket, get_parameter, IngestionReport
 from .ingestion.exceptions import UnspecifiedFormParameter, SubmissionFailure
 from .ingestion.processors import get_ingestion_processor
 from .inheritance_mode import InheritanceMode
@@ -325,89 +325,6 @@ def get_queue_manager(request, *, override_name):
     return (request.registry[INGESTION_QUEUE]
             if not override_name
             else IngestionQueueManager(request.registry, override_name=override_name))
-
-
-class IngestionError:
-    """
-    Holds info on an error that occurred in ingestion. Right now this consists of the
-    offending request body and the VCF row it occurred on.
-
-    This class doesn't really do much except specify the structure. It must align with that of FileProcessed
-    (reproduced as of 12/2/2020 below):
-
-        "file_ingestion_error": {
-            "title": "Ingestion Error Report",
-            "description": "This field is set when an error occurred in ingestion with all errors encountered",
-            "type": "array",
-            "items": {
-                "title": "Ingestion Error",
-                "type": "object",
-                "properties": {
-                    "body": {
-                        "type": "string",
-                        "index": false  # the intention is not to index this in the future
-                    },
-                    "row": {
-                        "type": "integer"
-                    }
-                }
-            }
-        }
-
-    """
-
-    def __init__(self, body, row):
-        self.body = body
-        self.row = row
-
-    def to_dict(self):
-        return {
-            'body': str(self.body),
-            'row': self.row
-        }
-
-
-class IngestionReport:
-    """
-    A "virtual" item on file_processed that contains detailed information on the ingestion run.
-    Not creating an item for this is a design decision. The output of this process is more for
-    debugging and not for auditing, so it does not merit an item at this time.
-    """
-    MAX_ERRORS = 100  # tune this to get more errors, 100 is a lot though and probably more than needed
-
-    def __init__(self):
-        self.grand_total = 0
-        self.errors = []
-
-    def brief_summary(self):
-        return ('INGESTION REPORT: There were %s total variants detected, of which %s were successful'
-                'and %s failed. Check ProcessedFile for full error output.' % (self.grand_total,
-                                                                              self.total_successful(),
-                                                                              self.total_errors()))
-
-    def total_successful(self):
-        return self.grand_total - len(self.errors)
-
-    def total_errors(self):
-        return len(self.errors)
-
-    def get_errors(self, limit=True):
-        """ Returns a limited number of errors, where limit can be True (self.MAX_ERRORS), False (no limit),
-            or an integer. """
-        if limit is True:
-            limit = self.MAX_ERRORS
-        elif limit is False:
-            limit = None
-        return self.errors[:limit]
-
-    def mark_success(self):
-        """ Marks the current row number as successful, which in this case just involves incrementing the total """
-        self.grand_total += 1
-
-    def mark_failure(self, *, body, row):
-        """ Marks the current row as failed, creating an IngestionError holding the response body and row. """
-        self.grand_total += 1
-        self.errors.append(IngestionError(body, row).to_dict())
 
 
 class IngestionQueueManager:
